Passing Day 08 - Part II :) (no need for an optimized way here - the naive one is sufficient O:) )

+++ 2025/day-08/src/part_02.rs
@@ -1,9 +1,9 @@
 use std::cmp::Ordering;
 
-pub fn resolve(s: &str) -> i32 {
+pub fn resolve(s: &str) -> i64 {
     let mut playground = transform_data(s);
     let (x, y) = playground.compute_until_single_circuit();
-    let final_result = x.x * y.x;
+    let final_result = (x.x as i64) * (y.x as i64);
     final_result
 }
